+++ core/lib/sn/SysNode.d.ts
@@ -1,14 +1,11 @@
-/// <reference types="node" />
 import { SysBase } from "./SysBase";
 import { IFn2Path, IConfig } from './CmnInterface';
-import m_fs = require('fs-extra');
 export declare class SysNode extends SysBase {
     protected readonly normalize: (src: string, _form: string) => string;
     loadPathAndVal(hPathFn2Exts: IFn2Path, fncLoaded: () => void, cfg: IConfig): void;
     protected readonly isApp: () => boolean;
-    readonly existsSync: typeof m_fs.existsSync;
-    readonly writeFile: typeof m_fs.writeFile;
     readonly savePic: (fn: string, data_url: string) => void;
-    readonly appendFile: typeof m_fs.appendFile;
+    readonly appendFile: any;
+    readonly ensureFileSync: (path: string) => any;
 }
 //# sourceMappingURL=SysNode.d.ts.map